Add inital demo/test code
This is mostly stuff ripped out of OpenSSL - in particular, aes_core.c
is an AES implementation written by Vincent Rijmen and others. I still
need to sort out the headers better, but compiling main.c and running it
on a DPU shows gibberish output for the ciphertext and the proper output
for the decrypted plaintext.