Add naiive performance measurement
This measures both wall time and DPU cycles, but the DPU cycle count
seems to be very off - with 17 tasklets, it measures 334 million
cycles to encrypt 64MB, which would come to a wall time of ~1.2s.
The real wall time is 17.04s.

Additionally, 17 tasklets gives a cyc/instr of 11.2, while 15 tasklets
gives ~2.2 (it's expected to be much higher because there are not enough
tasklets) but still has a longer wall clock time of 18.94s.

Back-of-the-envelope calculations suggest the hardware counter should
not overflow until at least 100s, but an overflow is the only
explanation I can come up with for this right now.