Replace hardcoded key with host-provided key
There's more work to be done on key management - I think this copies the
key directly to WRAM, which should be volatile and reasonably secure,
but it needs to be checked. It might also be possible to embed the key
in the DPU binary (e.g. hardcoded but changeable by the host), but it
still might be possible to copy it out from there.