Implement multi-DPU encryption
`dpu_AES_ecb` is now able to encrypt data on multiple DPUs at once, with
a specified number of DPUs (set at compile-time). However, the DPUs are
not aware of the size of the data and still encrypt/decrypt the entire
MRAM, so there is no reduction in the encryption time. The main result
is an increase in throughput for data larger than MRAM.