Add stricter input requirements
The length of the data is now subject to several requirements:
- it must be a multiple of 128
- it must be a multiple of the number of DPUs
- it must be a multiple of BOTH 128 and the number of DPUs

This means that the data can be split evenly among all the DPUs, with
the simple calculation that each DPU gets `length/nr_of_dpus` data. This
commit removes the mostly unchecked data-splitting code that existed before,
and fixes #1. It also adds helpful error messages and error-checking.