Refactor data transfer to be faster
Note that `in` can no longer be `const` because `dpu_prepare_xfer` may
result in modifying it (transfers to and from the DPU both use
`dpu_prepare_xfer`).

Also, the two `DPU_FOREACH` blocks could be condensed into one if `in ==
out` and the flag for `dpu_push_xfer` is `DPU_XFER_NO_RESET`.

Fixes #5