Add basic data output
This is the bare minimum setup for getting experimental data. It will
probably need an overhaul at some point, but I think it will work for
now.

There are measurements for eight sections of the code. It may turn out
that most of these are meaninglessly short and not variable, so the code
can be streamlined by removing them.

There are also two issues with this design, as it is:
- There is no way to get both instructions and cycles from the DPUs
- There is no way to do an operation without printing data (e.g. to
operate on data and then verify it by decrypting).

This means that getting IPC data and separate data for encryption and
decryption are going to be messy.