Set the DPU binary from the makefile
This ensures that the DPU binary loaded by the host is always the one
that was compiled for it. Hardcoding the path to the binary led to a bug
where tests running in experiment mode would compile a specific version
of the DPU binaries, but then the host would load whatever was most
recently compiled in regular build mode. This is because experiment mode
puts binaries in the experiment directory - not the build directory.