Fix segfault when using more than 64MB
As an artifact of when this only supported 1 DPU, the test data buffer
was statically allocated as the size of MRAM. So attempts to try using
more data (anything from 65MB with 2 DPUs to 40GB with 640 DPUs) would
overflow the buffer. The test data buffer is now allocated dynamically,
and has been tested with up to 40GB of data.