Fix segfault when operating on more than 64MB
As an artifact of when this supported only 1 DPU, the test data buffer
was statically allocated as the length of MRAM. This led to buffer
overflows when trying to operate on any amount of data larger than MRAM
(e.g. 64MB with 512 DPUs worked, but 65MB with 2 DPUs got a segfault).
The test data buffer is now allocated dynamically to match the length of
the test data, and has been tested up to 40GB.