Make inclusion of PIM-common headers more explicit
Both this repo and PIM-common have a common.h, so it was confusing to
tell which header a file was trying to include. This change means that
PIM-common headers have to be specified with a longer path - e.g.
`#include "PIM-common/common/include/common.h"` instead of `#include
"common.h"` - to remove the ambiguity and make it clear that the header is
from PIM-common.