Refactor main to include host execution mode
This allows the same executable to be used for executing on the DPU,
host, and (eventually) host with AES-NI; adds an option to
switch between encryption and decryption; and removes data validation.

Data validation should be reimplemented in a unit test in the future -
it just doesn't need to be done on every execution.

Implement un-accelerated host encryption