Add more compile-time parameters to the makefile
This adds `NOBULK` and `PERFCOUNT_TYPE` as compilation parameters which
allow the memory copy method (interleaved or non-interleaved) and
performance counter type (cycles or instructions) to be set during an
invocation of `make`. The performance counter type is included in the
output data.

It also brings out `TRANSFER_SIZE` into the makefile so it can be
configured the same way.

Finally, it fixes several issues with the makefile:
- Changes to header files did not trigger recompilation
- The experiment binary was always recompiled, even when nothing had
changed